Remove LICENSE
Some of the code in repository will be based on code provided by the
course, for which I have no copyright claim over and can't make any
guarantees about with regards to licensing.

So remove the LICENSE - I doubt this code will be of much use beyond my
own learning.